[Refactor] Moved utility test functions to a separate file.
